added styled components lib to json, restructured OV file structure, added basic style to root & app
