+++ client/src/overview/imageGallery/imageGallery_thumbnails.jsx
