+++ client/src/overview/productInformation/productInformation.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useProductsContext } from '../Context.jsx';
+import { useProductsContext } from '../../Context.jsx';
 
 export default function ProductInformation() {
   const { category, name, slogan } = useProductsContext();
